JWT Authentication completed

+++ backend/controllers/goalController.js
@@ -4,12 +4,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 // @desc Get goals 
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find() //finds all goals
+    const goals = await Goal.find({ user: req.user.id }) // user model. Access req.user with protect middleware, then match the id.
 
    res.status(200).json(goals) // sends back a json file as a response to the client.
 })
@@ -25,7 +27,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal) 
@@ -40,6 +43,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id) //user id from logged in user.
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) { // goals have a user field on it which is an object id. Turn it into a string before checking it.
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -58,6 +75,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id) //user id from logged in user.
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) { // goals have a user field on it which is an object id. Turn it into a string before checking it.
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
